+++ CHANGELOG.md
@@ -101,6 +101,8 @@
 
 * `dimred/densmap` component: Added a densMAP dimensionality reduction component (PR #748).
 
+* `annotete/scanvi` component: Added a component to annotate cells using scANVI (PR #833).
+
 * `transform/bpcells_regress_out` component: Added a component to regress out effects of confounding variables in the count matrix using BPCells (PR #863).
 
 * `transform/regress_out`: Allow providing 'input' and 'output' layers for scanpy regress_out functionality (PR #863).